% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sample_data.R
\docType{data}
\name{list_sample_data}
\alias{list_sample_data}
\title{simulation date set based on a SecondSpectrum Physical Split csv file}
\format{
a list of data.frames from a SecondSpectrum Physical Splits csv file. The data has been simulated and teams/players anonymised.
}
\source{
Data from a SecondSpectrum Physical Splits csv file
}
\usage{
list_sample_data
}
\description{
Simulation data to illustrate the function. Use the list_tidy_data function to transform the list of csv files into a tidy format for further analyses
}
\keyword{datasets}
