% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tidy_data.R
\name{list_tidy_data}
\alias{list_tidy_data}
\title{list_tidy data}
\usage{
list_tidy_data(x)
}
\arguments{
\item{x}{a list of SecondSpectrum Physical Splits csv files}
}
\description{
Creates a tidy data frame for further analyses
}
\details{
This function creates tidy data frames based on an a list of SecondSpectrum Physical Splits csv files. The output list has to be assigned to the GlobalEnvironment manually
}
\examples{
list_tidy_data(list_sample_data)
}
