% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{simulation date set based on a SecondSpectrum Physical Split csv file}
\format{
a data.frame from a SecondSpectrum Physical Splits csv file. The data has been simulated and teams/players anonymised.
}
\source{
Data from a SecondSpectrum Physical Splits csv file
}
\usage{
sample_data
}
\description{
Simulation data to illustrate the function. Use the tidy_data function to transform the csv file into a tidy format for further analyses
}
\keyword{datasets}
