% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{tidy data}
\usage{
tidy_data(x, wide_format = FALSE)
}
\arguments{
\item{x}{a SecondSpectrum Physical Splits csv file}
}
\description{
Creates a tidy dataframe for further analyses
}
\details{
This function creates a tidy data frame based on an a SecondSpectrum Physical Splits csv file. The output file has to be assigned to the GlobalEnvironment manually
}
\examples{
tidy_data(sample_data)
}
