% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpectrum.R
\docType{package}
\name{TidySpectrum}
\alias{TidySpectrum}
\alias{TidySpectrum-package}
\title{TudySpectrum: A package for tidying SecondSpectrum Physical Splits csv files}
\description{
This package takes a SecondSpectrum Physical Splits file/files, single files or as list, and outputs a tidy data frame/list of data frames in a tidy format for further analysis.
}
\author{
Lasse Ishøi \email{lasse.ishoei@gmail.com}
}
